Refactor: Organize project structure into feature-based folders
