--- app/src/main/java/com/ngatour/fruitclassifier/Data.kt
@@ -1,16 +1,0 @@
-package com.ngatour.fruitclassifier
-
-data class BatchEvaluationResult(
-    val total: Int,
-    val recognized: Int,
-    val unrecognized: Int,
-    val avgConfidence: Float,
-    val detailedResults: List<ClassificationResult>
-)
-data class SupabaseHistory(
-    val label: String,
-    val confidence: Float,
-    val description: String,
-    val timestamp: String,
-    val username: String
-)
