+++ app/src/main/java/com/ngatour/fruitclassifier/data/model/BatchEvaluationResult.kt
@@ -1,0 +1,9 @@
+package com.ngatour.fruitclassifier.data.model
+
+data class BatchEvaluationResult(
+    val total: Int,
+    val recognized: Int,
+    val unrecognized: Int,
+    val avgConfidence: Float,
+    val detailedResults: List<ClassificationResult>
+)
