+++ app/src/main/java/com/ngatour/fruitclassifier/data/model/SupabaseHistory.kt
@@ -1,0 +1,10 @@
+package com.ngatour.fruitclassifier.data.model
+
+data class SupabaseHistory(
+    val label: String,
+    val confidence: Float,
+    val description: String,
+    val timestamp: String,
+    val username: String
+)
+
