+++ app/src/main/java/com/ngatour/fruitclassifier/data/pref/UserPreferences.kt
@@ -1,4 +1,4 @@
-package com.ngatour.fruitclassifier
+package com.ngatour.fruitclassifier.data.pref
 
 import android.content.Context
 
@@ -16,5 +16,3 @@ class UserPreferences(context: Context) {
         prefs.edit().clear().apply()
     }
 }
-
-
