+++ app/src/main/java/com/ngatour/fruitclassifier/data/remote/SupabaseService.kt
@@ -1,9 +1,10 @@
-package com.ngatour.fruitclassifier
+package com.ngatour.fruitclassifier.data.remote
 
+import com.ngatour.fruitclassifier.data.model.SupabaseHistory
+import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.Header
 import retrofit2.http.POST
-import retrofit2.Response
 
 interface SupabaseService {
     @POST("history")
