+++ app/src/main/java/com/ngatour/fruitclassifier/ui/classify/ClassifyScreen.kt
@@ -1,16 +1,9 @@
-package com.ngatour.fruitclassifier
+package com.ngatour.fruitclassifier.ui.classify
 
-import android.content.Context
 import android.graphics.Bitmap
-import android.graphics.ImageDecoder
 import android.net.Uri
-import android.os.Build
-import android.provider.MediaStore
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import org.pytorch.IValue
-import org.pytorch.Module
-import org.pytorch.torchvision.TensorImageUtils
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
@@ -20,39 +13,30 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import coil.compose.rememberAsyncImagePainter
 import androidx.compose.ui.unit.dp
-import androidx.core.content.FileProvider
 import android.Manifest
-import android.content.Intent
-import android.graphics.BitmapFactory
-import android.graphics.ImageFormat
-import android.graphics.Paint
-import android.graphics.Rect
-import android.graphics.YuvImage
-import android.graphics.pdf.PdfDocument
 import android.widget.Toast
-import androidx.camera.core.ImageProxy
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.rememberPermissionState
 import androidx.compose.runtime.LaunchedEffect
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.google.accompanist.permissions.isGranted
+import com.ngatour.fruitclassifier.data.viewmodel.HistoryViewModel
+import com.ngatour.fruitclassifier.data.model.ClassificationResult
+import com.ngatour.fruitclassifier.util.classifyBitmap
 import com.ngatour.fruitclassifier.util.downloadModelFromUrl
+import com.ngatour.fruitclassifier.util.generatePDF
 import com.ngatour.fruitclassifier.util.isModelUpToDate
-import com.ngatour.fruitclassifier.util.modelFilePath
+import com.ngatour.fruitclassifier.util.saveBitmapToCache
+import com.ngatour.fruitclassifier.util.shareFile
+import com.ngatour.fruitclassifier.util.uriToBitmap
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import java.io.ByteArrayOutputStream
 import java.io.File
-import java.io.FileOutputStream
-import java.text.SimpleDateFormat
-import java.util.Date
-import java.util.Locale
-import kotlin.math.exp
 
 @OptIn(ExperimentalPermissionsApi::class,ExperimentalMaterial3Api::class)
 @Composable
-fun FruitClassifierApp(viewModel: HistoryViewModel) {
+fun FruitClassifierScreen(viewModel: HistoryViewModel) {
 
     val cameraPermissionState = rememberPermissionState(permission = Manifest.permission.CAMERA)
 
@@ -200,144 +184,4 @@ fun FruitClassifierApp(viewModel: HistoryViewModel) {
             }
         }
     }
-}
-
-fun uriToBitmap(context: Context, uri: Uri): Bitmap {
-    return if (Build.VERSION.SDK_INT < 28) {
-        MediaStore.Images.Media.getBitmap(context.contentResolver, uri)
-    } else {
-        val source = ImageDecoder.createSource(context.contentResolver, uri)
-        ImageDecoder.decodeBitmap(source)
-    }
-}
-
-fun ImageProxy.toBitmap(): Bitmap? {
-    val yBuffer = planes[0].buffer
-    val uBuffer = planes[1].buffer
-    val vBuffer = planes[2].buffer
-
-    val ySize = yBuffer.remaining()
-    val uSize = uBuffer.remaining()
-    val vSize = vBuffer.remaining()
-
-    val nv21 = ByteArray(ySize + uSize + vSize)
-    yBuffer.get(nv21, 0, ySize)
-    vBuffer.get(nv21, ySize, vSize)
-    uBuffer.get(nv21, ySize + vSize, uSize)
-
-    val yuvImage = YuvImage(nv21, ImageFormat.NV21, width, height, null)
-    val out = ByteArrayOutputStream()
-    yuvImage.compressToJpeg(Rect(0, 0, width, height), 100, out)
-    val imageBytes = out.toByteArray()
-    return BitmapFactory.decodeByteArray(imageBytes, 0, imageBytes.size)
-}
-
-fun saveBitmapToCache(context: Context, bitmap: Bitmap): Uri {
-    val file = File(context.cacheDir, "camera_image.jpg")
-    FileOutputStream(file).use { out ->
-        bitmap.compress(Bitmap.CompressFormat.JPEG, 100, out)
-    }
-    return FileProvider.getUriForFile(context, "${context.packageName}.provider", file)
-}
-
-fun classifyBitmap(context: Context, bitmap: Bitmap, modelName: String): ClassificationResult {
-    val labels = listOf("Banana", "Mango", "Orange", "Pineapple", "Salak")
-    val labelDescriptions = mapOf(
-        "Banana" to "Pisang adalah buah tropis yang kaya potasium dan vitamin B6.",
-        "Mango" to "Mangga memiliki rasa manis dan tekstur lembut, kaya akan vitamin C.",
-        "Orange" to "Jeruk merupakan sumber vitamin C, biasa dikonsumsi sebagai jus.",
-        "Pineapple" to "Nanas memiliki rasa asam-manis dan tinggi enzim bromelain.",
-        "Salak" to "Salak atau snake fruit memiliki rasa manis dan sedikit sepat."
-    )
-
-    val threshold = 0.75f // Minimum confidence that is considered valid
-
-    val module = Module.load(modelFilePath(context, "model_fruit_mobile.pt"))
-    val safeBitmap = convertToMutableBitmap(bitmap)
-    val resized = Bitmap.createScaledBitmap(safeBitmap, 224, 224, true)
-
-    val inputTensor = TensorImageUtils.bitmapToFloat32Tensor(
-        resized,
-        TensorImageUtils.TORCHVISION_NORM_MEAN_RGB,
-        TensorImageUtils.TORCHVISION_NORM_STD_RGB
-    )
-
-    val startTime = System.currentTimeMillis()
-    val outputTensor = module.forward(IValue.from(inputTensor)).toTensor()
-    val endTime = System.currentTimeMillis()
-    val duration = endTime - startTime
-
-    val rawScores = outputTensor.dataAsFloatArray
-    val probs = softmax(rawScores)
-    val maxIdx = probs.indices.maxByOrNull { probs[it] } ?: -1
-    val confidenceRaw = probs[maxIdx] // still within 0.0 - 1.0
-    val confidencePercent = confidenceRaw * 100
-
-    val dateFormat = SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS", Locale.getDefault())
-    val timestamp = dateFormat.format(Date())
-
-    return if (confidenceRaw >= threshold) {
-        val label = labels.getOrElse(maxIdx) { "Tidak Dikenal" }
-        val description = labelDescriptions[label] ?: "Tidak ada deskripsi."
-        ClassificationResult(label, confidencePercent, description, duration, timestamp)
-    } else {
-        ClassificationResult(
-            label = "Tidak dikenali",
-            confidence = confidencePercent,
-            description = "Gambar tidak sesuai dengan kelas buah yang telah dikenali.",
-            processTimeMs = duration,
-            timestamp = timestamp
-        )
-    }
-}
-
-fun convertToMutableBitmap(source: Bitmap): Bitmap {
-    return source.copy(Bitmap.Config.ARGB_8888, true)
-}
-
-fun softmax(logits: FloatArray): FloatArray {
-    val expScores = logits.map { exp(it.toDouble()) }
-    val sumExp = expScores.sum()
-    return expScores.map { (it / sumExp).toFloat() }.toFloatArray()
-}
-
-fun generatePDF(context: Context, result: ClassificationResult): File {
-    val pdfDoc = PdfDocument()
-    val pageInfo = PdfDocument.PageInfo.Builder(300, 400, 1).create()
-    val page = pdfDoc.startPage(pageInfo)
-    val canvas = page.canvas
-    val paint = Paint()
-
-    var y = 20
-    val spacing = 20
-
-    canvas.drawText("Fruit Classification Result", 10f, y.toFloat(), paint)
-    y += spacing
-    canvas.drawText("Label: ${result.label}", 10f, y.toFloat(), paint)
-    y += spacing
-    canvas.drawText("Confidence: ${"%.2f".format(result.confidence)}%", 10f, y.toFloat(), paint)
-    y += spacing
-    canvas.drawText("Time: ${result.timestamp}", 10f, y.toFloat(), paint)
-    y += spacing
-    canvas.drawText("Description:", 10f, y.toFloat(), paint)
-    y += spacing
-    canvas.drawText(result.description, 10f, y.toFloat(), paint)
-
-    pdfDoc.finishPage(page)
-
-    val file = File(context.getExternalFilesDir(null), "result_${System.currentTimeMillis()}.pdf")
-    pdfDoc.writeTo(FileOutputStream(file))
-    pdfDoc.close()
-
-    return file
-}
-
-fun shareFile(context: Context, file: File) {
-    val uri = FileProvider.getUriForFile(context, "${context.packageName}.provider", file)
-    val shareIntent = Intent(Intent.ACTION_SEND).apply {
-        type = "application/pdf"
-        putExtra(Intent.EXTRA_STREAM, uri)
-        addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-    }
-    context.startActivity(Intent.createChooser(shareIntent, "Bagikan hasil klasifikasi"))
 }
