+++ app/src/main/java/com/ngatour/fruitclassifier/ui/history/HistoryScreen.kt
@@ -1,4 +1,4 @@
-package com.ngatour.fruitclassifier
+package com.ngatour.fruitclassifier.ui.history
 
 import android.content.Intent
 import android.widget.Toast
@@ -14,6 +14,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.core.content.FileProvider
+import com.ngatour.fruitclassifier.data.viewmodel.HistoryViewModel
 
 @Composable
 fun HistoryScreen(viewModel: HistoryViewModel) {
