+++ app/src/main/java/com/ngatour/fruitclassifier/ui/nav/MainNavigation.kt
@@ -1,9 +1,8 @@
-package com.ngatour.fruitclassifier
+package com.ngatour.fruitclassifier.ui.nav
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
@@ -11,12 +10,17 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.compose.ui.Modifier
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
+import com.ngatour.fruitclassifier.ui.history.HistoryScreen
+import com.ngatour.fruitclassifier.data.viewmodel.HistoryViewModel
+import com.ngatour.fruitclassifier.ui.stats.StatsScreen
+import com.ngatour.fruitclassifier.ui.classify.FruitClassifierScreen
+import com.ngatour.fruitclassifier.ui.live.LiveCameraScreen
+import com.ngatour.fruitclassifier.ui.profile.ProfileScreen
 
 @OptIn(ExperimentalPermissionsApi::class)
 @Composable
 fun MainNavigation(isDarkMode: Boolean, onThemeToggle: () -> Unit) {
     val navController = rememberNavController()
-    val context = LocalContext.current
     val viewModel: HistoryViewModel = viewModel()
 
     Scaffold(
@@ -36,7 +40,7 @@ fun MainNavigation(isDarkMode: Boolean, onThemeToggle: () -> Unit) {
     ) { padding ->
         NavHost(navController, startDestination = Screen.Classify.route, Modifier.padding(padding)) {
             composable(Screen.Classify.route) {
-                FruitClassifierApp(viewModel = viewModel)
+                FruitClassifierScreen(viewModel = viewModel)
             }
             composable(Screen.History.route) {
                 HistoryScreen(viewModel = viewModel)
