+++ app/src/main/java/com/ngatour/fruitclassifier/ui/nav/Screen.kt
@@ -1,4 +1,4 @@
-package com.ngatour.fruitclassifier
+package com.ngatour.fruitclassifier.ui.nav
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Camera
